Fix inconsistencies in balance validation and trade route handling
Updates TradingController.php to validate leverage as a float, corrects redirect routes, and synchronizes leverage slider and modal functions in app/Views/trading/trade.php and app/Views/user/dashboard.php to handle decimal values accurately.

+++ app/Views/user/dashboard.php
@@ -912,7 +912,7 @@ function validateTradeAmount() {
 
 function showConfirmModal(action) {
     const amountInput = document.getElementById('tradeAmount');
-    const leverage = parseInt(document.getElementById('leverageValue').value) || 5;
+    const leverage = parseFloat(document.getElementById('leverageValue').value) || 5;
     const amount = parseFloat(amountInput.value) || 0;
     const balanceErrorDiv = document.getElementById('balanceError');
